Remove Default Formatter
Having a default formatter set by this function causes issues when you
have the formatter set elsewhere, and will cause it to use both at the
same time. In particular, if you set the formatter inside your
spec_helper.rb file.

Removing it will default the formatter to the one set in the any of the
users config files (.rspec, .rspec-local) and in a pinch, then can still
append this function call with --format=doc to get the desired format
they wish.